+++ OfficialCalculator/app/src/main/java/com/hmsecosystem/calculator/converter/UnitConverter.java
@@ -1,0 +1,56 @@
+/*
+      Copyright 2021. Futurewei Technologies Inc. All rights reserved.
+      Licensed under the Apache License, Version 2.0 (the "License");
+      you may not use this file except in compliance with the License.
+      You may obtain a copy of the License at
+        http:  www.apache.org/licenses/LICENSE-2.0
+      Unless required by applicable law or agreed to in writing, software
+      distributed under the License is distributed on an "AS IS" BASIS,
+      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+      See the License for the specific language governing permissions and
+      limitations under the License.
+*/
+package com.hmsecosystem.calculator.converter;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+import com.hmsecosystem.calculator.R;
+
+public class UnitConverter extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_unit_converter);
+    }
+
+    public void onClick(View v)
+    {
+        Intent i;
+        switch(v.getId())
+        {
+            case R.id.area:
+                i=new Intent(this,UnitArea.class);
+                startActivity(i);
+                break;
+            case R.id.length:
+                i=new Intent(this,UnitLength.class);
+                startActivity(i);
+                break;
+            case R.id.weight:
+                i=new Intent(this,UnitWeight.class);
+                startActivity(i);
+                break;
+            case R.id.tempearture:
+                i=new Intent(this,UnitTemperature.class);
+                startActivity(i);
+                break;
+        }
+    }
+
+}
