+++ OfficialCalculator/app/src/main/java/com/hmsecosystem/calculator/iap/Constants.java
@@ -1,0 +1,24 @@
+/*
+      Copyright 2021. Futurewei Technologies Inc. All rights reserved.
+      Licensed under the Apache License, Version 2.0 (the "License");
+      you may not use this file except in compliance with the License.
+      You may obtain a copy of the License at
+        http:  www.apache.org/licenses/LICENSE-2.0
+      Unless required by applicable law or agreed to in writing, software
+      distributed under the License is distributed on an "AS IS" BASIS,
+      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+      See the License for the specific language governing permissions and
+      limitations under the License.
+*/
+
+package com.hmsecosystem.calculator.iap;
+
+public class Constants {
+
+    /** RequestCode for pull up the pmsPay page */
+    public static final int REQ_CODE_BUY = 4002;
+
+    /** RequestCode for pull up the login page for isEnvReady interface */
+    public static final int REQ_CODE_LOGIN = 2001;
+
+}
